requirements-2.txt
iterative-stratification==0.1.6
numpy==1.17.3
Pillow==6.1.0
torch==1.4.0
torchvision==0.5.0
pandas==0.25.1
imgaug==0.3.0
scikit-learn==0.19.0
opencv-python==4.1.1.26
efficientnet-pytorch==0.5.0
tqdm==4.40.2
